Move InterestRates and IRate to domain
