+++ src/main/java/com/solvd/bankapp/domain/IRate.java
@@ -1,4 +1,4 @@
-package com.solvd.bankapp.service;
+package com.solvd.bankapp.domain;
 
 public interface IRate {
