+++ src/main/java/com/solvd/bankapp/domain/InterestRates.java
@@ -1,4 +1,4 @@
-package com.solvd.bankapp.service;
+package com.solvd.bankapp.domain;
 
 public enum InterestRates implements IRate {
 
@@ -16,7 +16,7 @@ public enum InterestRates implements IRate {
 
     @Override
     public double getRate() {
-        return 0;
+        return rate;
     }
 
     @Override
